refactor: 공통 로직 Handler로 분리
